perf: Add user configs to the cache check and remove the colors
Since the colors palette is generated from the defaults and the user
configurations (for the same version), there's no need to check the
colors.

The on_colors and terminal_colors fields of the user config (custom
functions) are dumped in a string to store them in the json cache. The
string data is used only to check if the config has change to use the
stored palette or create a new one, there's no need to decode them.

The on_highlights is used over the cache, so no need to add it to the
inputs.